removed Post folder and moved components to common and update imports
