+++ src/components/Feed/Feed.tsx
@@ -2,8 +2,8 @@ import Link from "next/link";
 import { useSession } from "next-auth/react";
 import { motion } from "framer-motion";
 import { trpc } from "@/utils/trpc";
-import PostSkeleton from "@/components/Post/PostSkeleton";
-import Post from "@/components/Post/Post";
+import PostSkeleton from "../common/PostSkeleton";
+import Post from "../common/Post";
 import UserCard from "@/components/Feed/UserCard";
 import useLikePost from "@/hooks/useLikePost";
 import useSavePost from "@/hooks/useSavePost";
