+++ src/components/common/Post.tsx
